requirements.txt
matplotlib==3.1.2
pandas==0.25.3
Pillow==9.2.0
seaborn==0.11.2
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
